Renamed the ConnectionsDecryptor to XmlConnectionsDecryptor and moved it into the Serializers namespace
